+++ apps/chat/tests/acceptance/features/step_definitions/postInvites_steps.js
@@ -1,8 +1,8 @@
-var CAM_MOCKS = require('../../../mock-data.js');
+var CAM_MOCKS = require('../../../common/mock-data.js');
 var SSTEPS = require('../../shared_steps.js');
 var SA_STEPS = require('../../sharedAPI_steps.js');
 
-var peersCollection_steps = module.exports = function(){
+module.exports = function(){
 
   /*
    *  Scenario: sending channel invitations via API
@@ -13,10 +13,10 @@ var peersCollection_steps = module.exports = function(){
   this.Given(/^a specified role is( not)? a subscriber to a channel$/,
     roleIsSubscriber);
   function roleIsSubscriber(arg1, next) {
-    this.roleId =
+    this.token =
       (arg1 === undefined)
-      ? CAM_MOCKS.roleId1
-      : CAM_MOCKS.nonsubscribedRoleId;
+      ? "12345" // role should be channel subscriber
+      : "88888" // role should not be
     next();
   }
 
@@ -37,8 +37,10 @@ var peersCollection_steps = module.exports = function(){
     var self = this;
     this.soproRequest("https://localhost/api/channels.invite", {
       method: "POST",
+      headers: {
+        "token-auth":this.token
+      },
       qs: {
-        role: this.roleId,
         channel: this.channel,
         user: this.peerId,
       },
@@ -78,7 +80,7 @@ var peersCollection_steps = module.exports = function(){
   this.Given(SA_STEPS.haveValidAuthToken.regex,
     SA_STEPS.haveValidAuthToken.fn);
 
-  this.Given(/^a specified role is( not)? a subscriber to a channel$/,
+  this.Given(/^the identity owning that token is( not)? a subscriber to a channel$/,
     roleIsSubscriber);
 
   this.Given(/^I have a peer who is( not)? a channel subscriber$/, 
@@ -105,7 +107,7 @@ var peersCollection_steps = module.exports = function(){
   this.Given(SA_STEPS.haveValidAuthToken.regex,
     SA_STEPS.haveValidAuthToken.fn);
 
-  this.Given(/^a specified role is( not)? a subscriber to a channel$/,
+  this.Given(/^the identity owning that token is( not)? a subscriber to a channel$/,
     roleIsSubscriber);
 
   this.Given(/^I have a peer$/, function(next){