+++ apps/chat/tests/acceptance/shared_steps.js
@@ -1,5 +1,5 @@
-var CAM_MOCKS = require('../mock-data.js');
-var protractorHelpers = require('../protractorHelpers.js')(browser,element);
+var CAM_MOCKS = require('../common/mock-data.js');
+var protractorHelpers = require('../common/protractor-helpers.js')(browser,element);
 var changeIdentity = protractorHelpers.changeIdentity;
 var fs = require('fs');
 
@@ -24,13 +24,22 @@ module.exports = {
       var roleIndex = (arg1 == undefined)? 0: 1;
       changeIdentity(roleIndex)
       .then(function(){
-        var channels = browser.element.all(by.css('#collection-channels .channel-item'));
-        channels.count()
-        .then(function (count) {
-          if (count >= 2) {
-            next();
+        element.all(by.css('#collection-channels > div.channel-collection > ul > li.sopro-config-collection-link.sopro-more-channels'))
+        .get(0)
+        .isDisplayed()
+        .then(function (isDisplayed) {
+          if (arg1) {
+            if (isDisplayed) {
+              next();
+            } else {
+              next.fail(new Error("Less than three Channels displayed"));
+            }
           } else {
-            next.fail(new Error("Less than two Channels displayed"));
+            if (isDisplayed) {
+              next.fail(new Error("More than two Channels displayed"));
+            } else {
+              next();
+            }
           }
         })
       })
@@ -56,7 +65,7 @@ module.exports = {
         }
       })
     }
-  },
+  },  
   roleChosen: {
     regex: /^I choose a( different)? role$/,
     fn: function (arg1, next) {